+++ packages/admin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @paljs/admin
 
+## 6.0.0
+
+### Major Changes
+
+- Upgrade Prisma Version to V5 and fix all breaking changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @paljs/types@6.0.0
+
 ## 5.3.3
 
 ### Patch Changes