+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @paljs/plugins
 
+## 6.0.0
+
+### Major Changes
+
+- Upgrade Prisma Version to V5 and fix all breaking changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @paljs/types@6.0.0
+  - @paljs/utils@6.0.0
+
 ## 5.3.3
 
 ### Patch Changes