+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @paljs/schema
 
+## 6.0.0
+
+### Major Changes
+
+- Upgrade Prisma Version to V5 and fix all breaking changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @paljs/display@6.0.0
+  - @paljs/types@6.0.0
+  - @paljs/utils@6.0.0
+
 ## 5.3.3
 
 ### Patch Changes